move auto-mapping to one folder
