+++ app/airflow/dags/libs/auto_mapping/core_rules_creation.py
